+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 - OAuth2 support ([#180](https://github.com/stac-utils/stac-asset/pull/180))
+- Retry configuration for HTTP ([#192](https://github.com/stac-utils/stac-asset/pull/192))
 
 ## [0.4.0] - 2024-06-19